requirements.txt
appnope==0.1.3
asn1crypto==1.5.1
asttokens==2.0.5
backcall==0.2.0
cffi==1.15.0
coincurve==17.0.0
configobj==5.0.6
cryptography==37.0.2
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
future==0.18.2
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
jupyter-client==7.3.1
jupyter-core==4.10.0
matplotlib-inline==0.1.3
mnemonic==0.20
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pysha3==1.0.2
python-dateutil==2.8.2
python-gnupg==0.4.8
python-ldap==3.4.0
pyzmq==22.3.0
secrets==1.0.2
six==1.16.0
stack-data==0.2.0
systematic==4.8.7
tornado==6.1
traitlets==5.2.0
wcwidth==0.2.5
